requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.9
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
attrs==24.2.0
    # via
    #   aiohttp
    #   cattrs
    #   jsonschema
    #   referencing
    #   requests-cache
beautifulsoup4==4.12.3
    # via wikipedia
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via
    #   google-auth
    #   streamlit
cattrs==24.1.2
    # via requests-cache
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via streamlit
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.7
    # via langchain-community
fastapi==0.115.6
    # via -r requirements.in
filetype==1.2.0
    # via langchain-google-genai
fonttools==4.55.1
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-ai-generativelanguage==0.6.10
    # via google-generativeai
google-api-core[grpc]==2.23.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.154.0
    # via google-generativeai
google-auth==2.36.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.3
    # via langchain-google-genai
googleapis-common-protos==1.66.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.68.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.68.1
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.0
    # via langsmith
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
inflect==7.4.0
    # via statsbombpy
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
joblib==1.4.2
    # via statsbombpy
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.7
    # via matplotlib
langchain==0.3.9
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.9
    # via -r requirements.in
langchain-core==0.3.21
    # via
    #   langchain
    #   langchain-community
    #   langchain-google-genai
    #   langchain-text-splitters
langchain-google-genai==2.0.6
    # via -r requirements.in
langchain-text-splitters==0.3.2
    # via langchain
langsmith==0.1.147
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.23.1
    # via dataclasses-json
matplotlib==3.9.3
    # via
    #   -r requirements.in
    #   mplsoccer
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.5.0
    # via inflect
mplsoccer==1.4.0
    # via -r requirements.in
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.15.2
    # via altair
numpy==1.26.4
    # via
    #   contourpy
    #   langchain
    #   langchain-community
    #   matplotlib
    #   mplsoccer
    #   pandas
    #   pydeck
    #   scipy
    #   seaborn
    #   streamlit
orjson==3.10.12
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   langchain-core
    #   marshmallow
    #   matplotlib
    #   streamlit
pandas==2.2.3
    # via
    #   mplsoccer
    #   seaborn
    #   statsbombpy
    #   streamlit
pillow==11.0.0
    # via
    #   matplotlib
    #   mplsoccer
    #   streamlit
platformdirs==4.3.6
    # via requests-cache
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.25.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==18.1.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.3
    # via
    #   fastapi
    #   google-generativeai
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-settings==2.6.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via
    #   httplib2
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   langchain
    #   langchain-community
    #   langsmith
    #   mplsoccer
    #   requests-cache
    #   requests-toolbelt
    #   statsbombpy
    #   streamlit
    #   wikipedia
requests-cache==1.2.1
    # via statsbombpy
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
scipy==1.14.1
    # via mplsoccer
seaborn==0.13.2
    # via mplsoccer
six==1.16.0
    # via
    #   python-dateutil
    #   url-normalize
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.36
    # via
    #   langchain
    #   langchain-community
starlette==0.41.3
    # via fastapi
statsbombpy==1.14.0
    # via -r requirements.in
streamlit==1.40.2
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via google-generativeai
typeguard==4.4.1
    # via inflect
typing-extensions==4.12.2
    # via
    #   altair
    #   fastapi
    #   google-generativeai
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typeguard
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
url-normalize==1.4.3
    # via requests-cache
urllib3==2.2.3
    # via
    #   requests
    #   requests-cache
wikipedia==1.4.0
    # via -r requirements.in
yarl==1.18.3
    # via aiohttp
